Fix issues causing UnicodeDecodeError
During installation, if the system's language is not using UTF-8 encoding Python will throw out a `UnicodeDecodeError` when reading the content from `README.rst` like this:

```
UnicodeDecodeError: 'ascii' codec can't decode byte 0xc2 in position 2881: ordinal not in range(128)
```

The problem is caused by a `\xc2\xa0` space after the "UploadedFileProperty" in section 0.1.2 of `README.rst`. This pull request removed the invisible space.